pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "universal-captcha-solver"
version = "0.1.0"
description = "A universal captcha solving library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0",
    "numpy>=1.24",
    "onnxruntime>=1.17",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
]
cli = [
    "typer>=0.12",
    "rich>=13.0",
]
selenium = [
    "selenium>=4.15",
]
playwright = [
    "playwright>=1.40",
]
gpu = [
    "onnxruntime-gpu>=1.17",
]
audio = [
    "soundfile>=0.12",
    "librosa>=0.10",
]
cloud = [
    "openai>=1.12",
    "anthropic>=0.25",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "mypy>=1.8",
]
all = [
    "universal-captcha-solver[server,cli,selenium,playwright,audio,cloud]",
]

[project.scripts]
captcha-solver = "captcha_solver.cli.app:app"

[project.entry-points."captcha_solver.solvers"]

[tool.hatch.build.targets.wheel]
packages = ["src/captcha_solver"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
